Implemented changes requested in pull request review
