+++ .github/workflows/UploadFiles_build_and_test_on_push.yml
@@ -11,14 +11,10 @@ on:
 
 jobs:
   build:
-    uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_test.yml@ISSUE-9
+    uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_build_test.yml@main
     with:
       workdir: Frends.SFTP.UploadFiles
-      prebuild_command: "docker run 
-      -v $PWD/Frends.SFTP.UploadFiles.Tests/Volumes/ssh_host_rsa_key.pub:/home/foo/.ssh/keys/ssh_host_rsa_key.pub:ro
-      -v $PWD/Frends.SFTP.UploadFiles.Tests/Volumes/share:/home/foo/share 
-      -p 2222:22 -d --name testikontti atmoz/sftp
-      foo:pass:::upload"
+      prebuild_command: docker-compose -f ./Frends.SFTP.UploadFiles.Tests/docker_compose.yml up -d
     secrets:
       badge_service_api_key: ${{ secrets.BADGE_SERVICE_API_KEY }}
       test_feed_api_key: ${{ secrets.TASKS_TEST_FEED_API_KEY }}
