+++ .github/workflows/UploadFiles_release.yml
@@ -5,14 +5,9 @@ on:
 
 jobs:
   build:
-    uses: FrendsPlatform/FrendsTasks/.github/workflows/linux_release.yml@ISSUE-9
+    uses: FrendsPlatform/FrendsTasks/.github/workflows/release.yml@main
     with:
       workdir: Frends.SFTP.UploadFiles
-      prebuild_command: "docker run 
-      -v $PWD/Frends.SFTP.UploadFiles.Tests/Volumes/ssh_host_rsa_key.pub:/home/foo/.ssh/keys/ssh_host_rsa_key.pub:ro
-      -v $PWD/Frends.SFTP.UploadFiles.Tests/Volumes/share:/home/foo/share 
-      -p 2222:22 -d --name testikontti atmoz/sftp
-      foo:pass:::upload"
 
     secrets:
       feed_api_key: ${{ secrets.TASKS_FEED_API_KEY }}
