--- Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/TestData/testfolder_2022/SFTPUploadTestFile(2) - Copy.txt
@@ -1,1 +1,0 @@
-This is a test file
