+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/TestData/testfolder_2022/SFTPUploadTestFile.txt
@@ -1,0 +1,1 @@
+This is a Test file
