+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/TestData/testfolder_2022/SFTPUploadTestFile2.txt
@@ -1,0 +1,1 @@
+This is another test file
