+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles.Tests/docker_compose.yml
@@ -1,0 +1,16 @@
+﻿version: '3'
+
+# Usage https://github.com/atmoz/sftp
+
+services:
+ ftpd_server:
+  image: atmoz/sftp
+  container_name: testikontti
+  ports:
+   - "2222:22"
+  volumes:
+   - ../Volumes/ssh_host_rsa_key.pub:/home/foo/.ssh/keys/ssh_host_rsa_key.pub:ro
+   - ../Volumes/share:/home/foo/share
+  command: 
+   - "foo:pass:::upload"
+
