+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/Destination.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace Frends.SFTP.UploadFiles.Definitions
@@ -66,7 +65,7 @@ namespace Frends.SFTP.UploadFiles.Definitions
         /// <summary>
         /// File encoding to be used. A partial list of possible encodings: https://en.wikipedia.org/wiki/Windows_code_page#List.
         /// </summary>
-        /// <example>"utf-64"</example>
+        /// <example>utf-64</example>
         [UIHint(nameof(FileContentEncoding), "", FileEncoding.Other)]
         public string FileContentEncodingInString { get; set; }
