+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/EndPoint.cs
@@ -1,12 +1,6 @@
-﻿// Self-explanatory enum
-#pragma warning disable 1591
-
-namespace Frends.SFTP.UploadFiles.Definitions
+﻿namespace Frends.SFTP.UploadFiles.Definitions
 {
-    /// <summary>
-    /// Defines EndPoint of the transfer. This is used in methods which can be used for both directions.
-    /// </summary>
-    public enum EndPoint
+    internal enum EndPoint
     {
         Source,
         Destination
