+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/TransferResult.cs
@@ -1,17 +1,8 @@
 ﻿namespace Frends.SFTP.UploadFiles.Definitions
 {
-    /// <summary>
-    /// Enumeration for transfer result. Used in SFTPLogger and operations log.
-    /// </summary>
-    public enum TransferResult
+    internal enum TransferResult
     {
-        /// <summary>
-        /// Defines successful transfer.
-        /// </summary>
         Success,
-        /// <summary>
-        /// Defines failed transfer.
-        /// </summary>
         Fail
     }
 }
