+++ Frends.SFTP.UploadFiles/Volumes/ssh_host_rsa_key.pub
@@ -1,0 +1,8 @@
+---- BEGIN SSH2 PUBLIC KEY ----
+AAAAB3NzaC1yc2EAAAADAQABAAABAQC86JmeYefZ56B8nQE7gj4bs9q7QMJuXdsI
+1cvR1hFjiTVo+yy+PK7Eo5FUKNIC+IWPYz6hn46M9Jgva4sAZQYr3ocLu8+jEm+O
+QrVrjrlD9IPnwf/EhLGs6uQpzr4VsNTPOyD+PKGGlPkdMTc5h026CB0WXLuOvH0d
+myV3+AVlzoA+1RtM9t/948WGEP7iPblZc/6jr72ef+OCGNw4txcpqvyd4M4wHUDw
+b6T/3xppzO7hJb0Kwk20lxyK9Oe2UEW4roCj1if/qWXWM/2Mm6MkrRZCZIRmPZfm
+SkHIrmkMo8yS1rDheABvc2R3PuryEMmh/gG9MBBLFB17ZGLuOVr3
+---- END SSH2 PUBLIC KEY ----
